add: hooks method
