+++ index.js
@@ -26,9 +26,8 @@ const ajvFormats = require('ajv-formats')
 const { unset: lunset } = require('lodash')
 const { readdirSync } = require('fs')
 const { join } = require('path')
-const { ObjectId } = require('mongodb')
 const { JSONPath } = require('jsonpath-plus')
-const { hrtime } = require('node:process')
+
 
 const myPackage = require('./package')
 const fastifyEnvSchema = require('./envSchema')
@@ -40,21 +39,14 @@ const joinPlugin = require('./lib/joinPlugin')
 const { castCollectionId } = require('./lib/pkFactories')
 const {
   SCHEMA_CUSTOM_KEYWORDS,
-  SETCMD,
-  PUSHCMD,
-  PULLCMD,
-  UNSETCMD,
-  ADDTOSETCMD,
-  QUERY,
-  RAW_PROJECTION,
-  INVALID_USERID,
 } = require('./lib/consts')
 const { registerMongoInstances } = require('./lib/mongo/mongo-plugin')
-const { getAjvResponseValidationFunction, ajvSerializer, ajvValidator } = require('./lib/validatorGetters')
+const { ajvSerializer, ajvValidator } = require('./lib/validatorGetters')
 const { pointerSeparator } = require('./lib/JSONPath.utils')
 const { registerHelperRoutes } = require('./lib/helpersRoutes')
 const AdditionalCaster = require('./lib/AdditionalCaster')
 const { addValidatorCompiler, addSerializerCompiler } = require('./lib/compilers')
+const { addAclHook, addHealthHooks, addPreHandlerHooks, addLookupHook, addPatchViewHook, addModelNameHook } = require('./lib/hooks')
 
 function decorateCrud(fastify, model, modelName) {
   fastify.decorate('crudService', model.crudService)
@@ -63,9 +55,7 @@ function decorateCrud(fastify, model, modelName) {
   fastify.decorate('jsonSchemaGenerator', model.jsonSchemaGenerator)
   fastify.decorate('jsonSchemaGeneratorWithNested', model.jsonSchemaGenerator)
   fastify.decorate('modelName', modelName)
-  fastify.addHook('onRequest', async(request) => {
-    request.modelName = modelName
-  })
+  addModelNameHook(fastify, modelName)
 }
 
 async function registerCrud(fastify, { modelName, isView }) {
@@ -93,17 +83,6 @@ async function registerViewCrud(fastify, { modelName }) {
 
   decorateCrud(fastify, viewDependencies, modelName)
   await fastify.register(httpInterface, { prefix, registerGetters: false, registerSetters: true })
-}
-
-function mapLookupToObjectId(reference) {
-  // consider both one-to-one and one-to-many relationships
-  if (Array.isArray(reference)) {
-    return reference
-      .map(ref => (ref?.value ? new ObjectId(ref.value) : undefined))
-      .filter(Boolean)
-  }
-
-  return reference?.value ? new ObjectId(reference.value) : undefined
 }
 
 async function registerViewCrudLookup(fastify, { modelName, lookupModel }) {
@@ -228,82 +207,17 @@ async function setupCruds(fastify) {
     await fastify.register(registerHelperRoutes, { prefix: HELPERS_PREFIX })
   }
 
-  /** --------------------------  HOOKS ----------------------------------- */
-
-  // To allow writing views without having to rewrite all the logic of the HttpInterface,
-  // it was decided to adapt the fields of the calls towards the view by converting them
-  // to the fields of the underlying collection, thus hiding the complexity on the client
-  // side while maintaining consistent interfaces.
-  // This assumes that the key of the value is in the field "value" and should be made configurable.
-  const lookups = fastify.models ? Object.entries(fastify.models)
-    .reduce((accumulator, [modelName, { viewDependencies }]) => {
-      const viewLookups = viewDependencies.lookupsModels?.map(({ lookup }) => lookup) || []
-      accumulator[modelName] = viewLookups
-      return accumulator
-    }, {}) : {}
-  fastify.addHook('preHandler', (request, _reply, done) => {
-    const modelLookups = lookups[request.modelName] || []
-    for (const { as, localField } of modelLookups) {
-      if (request?.body?.[as]) {
-        const lookupReference = request.body[as]
-        delete request.body[as]
-
-        request.body[localField] = mapLookupToObjectId(lookupReference)
-      }
-
-      for (const command of [SETCMD, UNSETCMD, PUSHCMD, PULLCMD, ADDTOSETCMD]) {
-        if (request?.body?.[command]?.[as]) {
-          const lookupReference = request.body[command][as]
-          delete request.body[command][as]
-
-          request.body[command][localField] = mapLookupToObjectId(lookupReference)
-        }
-      }
-    }
-
-    done()
-  })
-
-  // To obtain the updated object with a consistent interface after a patch,
-  // it is necessary to retrieve the view object again before returning it to the client.
-  fastify.addHook('preSerialization', async function preSerializer(request, _reply, payload) {
-    const { _id } = payload || {}
-
-    if (request.method === 'PATCH' && _id && fastify.models) {
-      const { crudService, isView } = fastify.models[request.modelName] || {}
-      if (isView) {
-        const docId = this.castCollectionId(_id)
-        // eslint-disable-next-line no-underscore-dangle
-        const doc = await crudService._mongoCollection.findOne({ _id: docId })
-        const validatePatch = getAjvResponseValidationFunction(request.routeSchema.response['200'])
-        validatePatch(doc)
-        return doc
-      }
-    }
-    return payload
-  })
+  /** ===============================  HOOKS  ===============================  */
+  addLookupHook(fastify)
+  addPatchViewHook(fastify)
 }
 
 /* =============================================================================== */
 
 module.exports = async function plugin(fastify, opts) {
-  const start = hrtime.bigint()
-
-  fastify.addHook('onReady', () => {
-    fastify.log.info({ elapsedMs: Number(hrtime.bigint() - start) / 1_000_000 }, 'ready event reached')
-  })
-  fastify.addHook('onListen', () => {
-    fastify.log.info({ elapsedMs: Number(hrtime.bigint() - start) / 1_000_000 }, 'listen event reached')
-  })
-
-  fastify.addHook('onRequest', async(request) => {
-    if (request.headers.acl_rows) {
-      request.headers.acl_rows = JSON.parse(request.headers.acl_rows)
-    }
-  })
-
-  fastify.addHook('preHandler', injectContextInRequest)
-  fastify.addHook('preHandler', request => parseEncodedJsonQueryParams(fastify.log, request))
+  addHealthHooks(fastify)
+  addAclHook(fastify)
+  addPreHandlerHooks(fastify)
 
   await fastify.register(fastifyEnv, { schema: fastifyEnvSchema, data: opts })
   await fastify.register(fastifyMultipart, {
@@ -426,44 +340,6 @@ async function readinessHandler(fastify) {
   const statusOK = await isMongoUp(fastify)
 
   return { statusOK }
-}
-
-async function parseEncodedJsonQueryParams(logger, request) {
-  if (request.headers.json_query_params_encoding) {
-    logger.warn('You\'re using the json_query_params_encoding header but it\'s deprecated and its support is going to be dropped in the next major release. Use json-query-params-encoding instead.')
-  }
-
-  // TODO remove request.headers.json_query_params_encoding fallback in v7.0.0
-  const jsonQueryParamsEncoding = request.headers['json-query-params-encoding'] || request.headers.json_query_params_encoding
-  switch (jsonQueryParamsEncoding) {
-  case 'base64': {
-    const queryJsonFields = [QUERY, RAW_PROJECTION]
-    for (const field of queryJsonFields) {
-      if (request.query[field]) {
-        request.query[field] = Buffer.from(request.query[field], jsonQueryParamsEncoding).toString()
-      }
-    }
-    break
-  }
-  default: break
-  }
-}
-
-async function injectContextInRequest(request) {
-  const userIdHeader = request.headers[this.userIdHeaderKey]
-  const isUserHeaderInvalid = INVALID_USERID.includes(userIdHeader)
-
-  let userId = 'public'
-
-  if (userIdHeader && !isUserHeaderInvalid) {
-    userId = userIdHeader
-  }
-
-  request.crudContext = {
-    log: request.log,
-    userId,
-    now: new Date(),
-  }
 }
 
 module.exports.readinessHandler = readinessHandler
